requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
biopython>=1.79
requests>=2.26.0
tqdm>=4.62.0
xgboost>=1.5.0
lightgbm>=3.3.0
tensorflow>=2.8.0
torch>=1.10.0
optuna>=2.10.0
joblib>=1.1.0 